Safari-611.1.10/LayoutTests/contentfiltering/allow-after-add-data-expected.txt

main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame
frame "<!--frame1-->" - willPerformClientRedirectToURL: resources/pass.html
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
frame "<!--frame1-->" - didCancelClientRedirectForFrame
frame "<!--frame1-->" - didCommitLoadForFrame
frame "<!--frame1-->" - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFinishLoadForFrame


--------
Frame: '<!--frame1-->'
--------
PASS

============== Back Forward List ==============
curr->  (file test):contentfiltering/allow-after-add-data.html  **nav target**
            (file test):contentfiltering/resources/pass.html (in frame "<!--frame1-->")
===============================================
